backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
geoalchemy2
shapely
pydantic-settings
python-dotenv
python-jose[cryptography]
bcrypt
python-multipart
arq
redis
osmnx
networkx
